+++ web/src/widget/custom_button.dart
@@ -1,0 +1,45 @@
+//Copyright (C) 2012 Sergey Akopkokhyants. All Rights Reserved.
+//Author: akserg
+
+library summary;
+
+import 'dart:html' as dart_html;
+
+import 'package:dart_web_toolkit/event.dart' as event;
+import 'package:dart_web_toolkit/ui.dart' as ui;
+
+void main() {
+
+  ui.VerticalPanel vpanel = new ui.VerticalPanel();
+
+  ui.HorizontalPanel pushPanel = new ui.HorizontalPanel();
+  pushPanel.spacing = 10;
+
+  ui.HorizontalPanel togglePanel = new ui.HorizontalPanel();
+  togglePanel.spacing = 10;
+
+  // Combine all the panels
+  vpanel.add(new ui.Html("Custom Button"));
+  vpanel.add(pushPanel);
+  vpanel.add(new ui.Html("<br><br>PushButtons and ToggleButtons allow you to customize the look of your buttons"));
+  vpanel.add(togglePanel);
+
+  ui.PushButton normalPushButton = new ui.PushButton.fromImage(new ui.Image("img/darts.jpg"));
+  pushPanel.add(normalPushButton);
+
+  // Add a disabled PushButton
+  ui.PushButton disabledPushButton = new ui.PushButton.fromImage(new ui.Image("img/darts.jpg"));
+  disabledPushButton.enabled = false;
+  pushPanel.add(disabledPushButton);
+
+  // Add a normal ToggleButton
+  ui.ToggleButton normalToggleButton = new ui.ToggleButton.fromImage(new ui.Image("img/darts.jpg"));
+  togglePanel.add(normalToggleButton);
+
+  // Add a disabled ToggleButton
+  ui.ToggleButton disabledToggleButton = new ui.ToggleButton.fromImage(new ui.Image("img/darts.jpg"));
+  disabledToggleButton.enabled = false;
+  togglePanel.add(disabledToggleButton);
+
+  ui.RootPanel.get().add(vpanel);
+}