+++ web/src/widget/html.dart
@@ -1,0 +1,15 @@
+//Copyright (C) 2012 Sergey Akopkokhyants. All Rights Reserved.
+//Author: akserg
+
+library summary;
+
+import 'dart:html' as dart_html;
+
+import 'package:dart_web_toolkit/event.dart' as event;
+import 'package:dart_web_toolkit/ui.dart' as ui;
+
+void main() {
+
+  ui.Html html = new ui.Html("<div id='dred' style='background-color: yellow; border: 1px dotted red; width: 200px; text-align: center;'> This is an HTML Widget </div>");
+  ui.RootPanel.get().add(html);
+}