+++ web/src/widget/image.dart
@@ -1,0 +1,14 @@
+//Copyright (C) 2012 Sergey Akopkokhyants. All Rights Reserved.
+//Author: akserg
+
+library summary;
+
+import 'dart:html' as dart_html;
+
+import 'package:dart_web_toolkit/event.dart' as event;
+import 'package:dart_web_toolkit/ui.dart' as ui;
+
+void main() {
+  ui.Image image = new ui.Image("img/darts.jpg");
+  ui.RootPanel.get().add(image);
+}